nvim: Add virtualedit=block
This allows me to use visual block mode to select outside the bounds of
the file. However when yanking, vim.highlight.on_yank() would attempt to
highlight text that doesn't exist (it gave an index out of bounds
error), so we wrap it in a pcall() to suppress the errors.