nvim: Don't auto insert comments on o and O
I tried to use this for a while, but I find this gets in the way more
than it helps. Especially having comments with no surrounding blank
lines (e.g. at the start/end of a file) and you have to do a song and
dance to insert a blank line lol. The r option easily lets me continue
comments if I want to without getting in the way so I'm just going to
use that.