nvim: Fix format on save erroneously clearing autogroup when attaching
This meant that the autocommand wouldn't be present in the case where
you open a buffer, then open a different buffer and then return to the
first buffer. This is because opening the second buffer cleared the
autogroup and returning to the first buffer again doesn't trigger
LspAttach (the LSP is already attached).

We fix this by setting clear = false on the autogroup definition and
then making sure exactly one formatting autocommand is present for each
buffer. I don't think this check is entirely necessary as the logic
earlier in the function ensures we have one formatter per buffer anyway,
but I think this is a good failsafe to have.