nvim: Add Windows config
In order for nvim on Windows to find my config, we symlink
`AppData/Local/nvim/` to `.config/nvim/`.

However there were some Windows specific issues that this commit fixes:

* `chezmoi.lua`'s source path mixed backslashes and forward slashes (we
  use vim.fs.normalize instead now).
* We set the default shell to powershell instead of cmd
* Packer attempted to create a directory with the full URL of
  clang_extensions as the name, which didn't work, so we alias it.

This isn't perfect yet, as there are issues with my git plugins (neogit
definitely doesn't play nice with my scoop installation of git), but
it's a start.