wezterm: Add shortcuts for typing Alt/Ctrl-h/j/k/l
These shortcuts are now used by the smart-splits.nvim integration. But
it is still useful to be able to type them (e.g. Ctrl-L for clearing the
screen, or Ctrl-J/K for moving commits in lazygit). So, add the shortcuts
back as leader key shortcuts, similar to the existing Ctrl-Space leader
key mapping.