wezterm: Disable Commit Mono coding ligatures
There isn't a ligature for the Haskell bind operator (`>>=`, AKA right
shift assign), so it is rendered as a `>` followed by the `>=` ligature,
which looks ugly. The creator of Commit Mono says that these ligatures
are "low priority" in [this issue](https://github.com/eigilnikolajsen/commit-mono/issues/64),
so it looks like this won't be fixed.