chezmoi: Use "pwsh.exe" instead of full executable path
On WSL NixOS with services.envfs enabled, the full path of pwsh.exe (and
other executables) returned by `lookPath` is spoofed to a non-existent
path under `/usr/bin/` to ensure compatibility. If we use "pwsh.exe" as
the executable (instead using `lookPath` as a boolean) then NixOS will
be able to find it again.